Rename middlewares::jwt -> middlewares::auth.
