+++ src/server/middlewares.rs
@@ -1,1 +1,1 @@
-pub mod jwt;
+pub mod auth;
