+++ src/server/middlewares/auth.rs
